+++ credentials/LinklyOAuth2Api.credentials.ts
@@ -1,0 +1,62 @@
+import type {
+	ICredentialType,
+	INodeProperties,
+} from 'n8n-workflow';
+
+export class LinklyOAuth2Api implements ICredentialType {
+	name = 'linklyOAuth2Api';
+	displayName = 'Linkly OAuth2 API';
+	documentationUrl = 'https://linklyhq.com/support/api';
+	extends = ['oAuth2Api'];
+
+	properties: INodeProperties[] = [
+		{
+			displayName: 'Grant Type',
+			name: 'grantType',
+			type: 'hidden',
+			default: 'authorizationCode',
+		},
+		{
+			displayName: 'Authorization URL',
+			name: 'authUrl',
+			type: 'hidden',
+			default: 'https://app.linklyhq.com/oauth/authorize',
+		},
+		{
+			displayName: 'Access Token URL',
+			name: 'accessTokenUrl',
+			type: 'hidden',
+			default: 'https://app.linklyhq.com/oauth/token',
+		},
+		{
+			displayName: 'Client ID',
+			name: 'clientId',
+			type: 'hidden',
+			default: 'linkly_n8n_25316062113802ac',
+		},
+		{
+			displayName: 'Client Secret',
+			name: 'clientSecret',
+			type: 'hidden',
+			default: '3ecREEKyGHGnTaZlt_HLCEeKNsm97G2kqgJOPVAS',
+		},
+		{
+			displayName: 'Authentication',
+			name: 'authentication',
+			type: 'hidden',
+			default: 'body',
+		},
+		{
+			displayName: 'Auth URI Query Parameters',
+			name: 'authQueryParameters',
+			type: 'hidden',
+			default: '',
+		},
+		{
+			displayName: 'Scope',
+			name: 'scope',
+			type: 'hidden',
+			default: '',
+		},
+	];
+}